cleaned up folder system
