+++ FrontEnd/src/App.js
@@ -1,8 +1,34 @@
-import './styles/App.css'
-import React from 'react'
+import './styles/App.css';
+import React, {useState} from 'react';
+import { Routes, Route} from 'react-router-dom'
+import Home from './pages/Home'
+import Navbar from './components/Navbar'
+import Parks from './pages/Parks';
+import About from './pages/About'
+import Search from './components/Search'
 
-const App = () => {
-    return()
+
+
+
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <p>
+          Edit <code>src/App.js</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header>
+    </div>
+  );
 }
 
-export default App
+export default App;
