+++ FrontEnd/src/Index.js
@@ -1,5 +1,16 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import './styles/Index.css'
-import App from './App'
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './styles/index.css';
+import App from './App';
+import {BrowserRouter} from 'react-router-dom'
 
+
+
+ReactDOM.render(
+  <React.StrictMode>
+    <BrowserRouter>
+    <App />
+    </BrowserRouter>
+  </React.StrictMode>,
+  document.getElementById('root')
+);
