--- FrontEnd/src/client/src/index.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './styles/index.css';
-import App from './App';
-import {BrowserRouter} from 'react-router-dom'
-
-
-
-ReactDOM.render(
-  <React.StrictMode>
-    <BrowserRouter>
-    <App />
-    </BrowserRouter>
-  </React.StrictMode>,
-  document.getElementById('root')
-);
