--- FrontEnd/src/components/Park.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const Park = (props) => {
-    return()
-}
-
-export default park
