--- FrontEnd/src/components/Ride.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const Ride = (props) => {
-    return()
-}
-
-export default Ride
